refactor: move exported type definitions to declaration file
