+++ packages/docusaurus-plugin-client-redirects/src/__tests__/collectRedirects.test.ts
@@ -5,13 +5,14 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import type {PluginContext, UserPluginOptions} from '../types';
+import type {PluginContext} from '../types';
 import collectRedirects from '../collectRedirects';
 import normalizePluginOptions from '../normalizePluginOptions';
 import {removeTrailingSlash} from '@docusaurus/utils';
+import type {Options} from '@docusaurus/plugin-client-redirects';
 
 function createTestPluginContext(
-  options?: UserPluginOptions,
+  options?: Options,
   relativeRoutesPaths: string[] = [],
 ): PluginContext {
   return {
