+++ packages/docusaurus-plugin-client-redirects/src/index.ts
@@ -6,7 +6,8 @@
  */
 
 import type {LoadContext, Plugin, Props} from '@docusaurus/types';
-import type {UserPluginOptions, PluginContext, RedirectMetadata} from './types';
+import type {PluginContext, RedirectMetadata} from './types';
+import type {PluginOptions} from '@docusaurus/plugin-client-redirects';
 
 import normalizePluginOptions from './normalizePluginOptions';
 import collectRedirects from './collectRedirects';
@@ -18,7 +19,7 @@ import {removePrefix, addLeadingSlash} from '@docusaurus/utils';
 
 export default function pluginClientRedirectsPages(
   context: LoadContext,
-  opts: UserPluginOptions,
+  opts: PluginOptions,
 ): Plugin<unknown> {
   const {trailingSlash} = context.siteConfig;
