+++ packages/docusaurus-plugin-client-redirects/src/normalizePluginOptions.ts
@@ -5,7 +5,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import type {PluginOptions, RedirectOption, UserPluginOptions} from './types';
+import type {
+  PluginOptions,
+  Options as UserPluginOptions,
+} from '@docusaurus/plugin-client-redirects';
+import type {RedirectOption} from './types';
 import {Joi, PathnameSchema} from '@docusaurus/utils-validation';
 import {DEFAULT_PLUGIN_ID} from '@docusaurus/utils';
