+++ packages/docusaurus-plugin-content-docs/src/__tests__/cli.test.ts
@@ -7,7 +7,10 @@
 
 import path from 'path';
 import {cliDocsVersionCommand} from '../cli';
-import type {PathOptions, SidebarOptions} from '../types';
+import type {
+  PathOptions,
+  SidebarOptions,
+} from '@docusaurus/plugin-content-docs';
 import fs from 'fs-extra';
 import {
   getVersionedDocsDirPath,
