+++ packages/docusaurus-plugin-content-docs/src/__tests__/docs.test.ts
@@ -19,12 +19,14 @@ import {readVersionsMetadata} from '../versions';
 import type {
   DocFile,
   DocMetadataBase,
-  MetadataOptions,
   VersionMetadata,
-  PluginOptions,
-  EditUrlFunction,
   DocNavLink,
 } from '../types';
+import type {
+  MetadataOptions,
+  PluginOptions,
+  EditUrlFunction,
+} from '@docusaurus/plugin-content-docs';
 import type {LoadContext} from '@docusaurus/types';
 import {DEFAULT_OPTIONS} from '../options';
 import type {Optional} from 'utility-types';
