+++ packages/docusaurus-plugin-content-docs/src/__tests__/versions.test.ts
@@ -14,8 +14,9 @@ import {
 } from '../versions';
 import {DEFAULT_OPTIONS} from '../options';
 import {DEFAULT_PLUGIN_ID} from '@docusaurus/utils';
-import type {PluginOptions, VersionMetadata} from '../types';
+import type {VersionMetadata} from '../types';
 import type {I18n} from '@docusaurus/types';
+import type {PluginOptions} from '@docusaurus/plugin-content-docs';
 
 const DefaultI18N: I18n = {
   currentLocale: 'en',
