+++ packages/docusaurus-plugin-content-docs/src/sidebars/types.ts
@@ -6,12 +6,11 @@
  */
 
 import type {Optional, Required} from 'utility-types';
+import type {DocMetadataBase, VersionMetadata} from '../types';
 import type {
-  DocMetadataBase,
-  VersionMetadata,
   NumberPrefixParser,
   SidebarOptions,
-} from '../types';
+} from '@docusaurus/plugin-content-docs';
 
 // Makes all properties visible when hovering over the type
 type Expand<T extends Record<string, unknown>> = {[P in keyof T]: T[P]};
