+++ packages/docusaurus-preset-classic/src/preset-classic.d.ts
@@ -5,18 +5,28 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import type {Options as DocsPluginOptions} from '@docusaurus/plugin-content-docs';
+import type {Options as BlogPluginOptions} from '@docusaurus/plugin-content-blog';
+import type {Options as PagesPluginOptions} from '@docusaurus/plugin-content-pages';
+import type {Options as SitemapPluginOptions} from '@docusaurus/plugin-sitemap';
+import type {Options as GAPluginOptions} from '@docusaurus/plugin-google-analytics';
+import type {Options as GtagPluginOptions} from '@docusaurus/plugin-google-gtag';
+import type {Options as ThemeOptions} from '@docusaurus/theme-classic';
+import type {ThemeConfig as BaseThemeConfig} from '@docusaurus/types';
+import type {UserThemeConfig as ClassicThemeConfig} from '@docusaurus/theme-common';
+import type {UserThemeConfig as AlgoliaThemeConfig} from '@docusaurus/theme-search-algolia';
+
 export type Options = {
   debug?: boolean;
-  docs?: false | import('@docusaurus/plugin-content-docs').Options;
-  blog?: false | import('@docusaurus/plugin-content-blog').Options;
-  pages?: false | import('@docusaurus/plugin-content-pages').Options;
-  sitemap?: false | import('@docusaurus/plugin-sitemap').Options;
-  theme?: import('@docusaurus/theme-classic').Options;
-  googleAnalytics?: import('@docusaurus/plugin-google-analytics').Options;
-  gtag?: import('@docusaurus/plugin-google-gtag').Options;
+  docs?: false | DocsPluginOptions;
+  blog?: false | BlogPluginOptions;
+  pages?: false | PagesPluginOptions;
+  sitemap?: false | SitemapPluginOptions;
+  theme?: ThemeOptions;
+  googleAnalytics?: GAPluginOptions;
+  gtag?: GtagPluginOptions;
 };
 
-export type ThemeConfig = import('@docusaurus/types').ThemeConfig &
-  import('@docusaurus/theme-common').UserThemeConfig & {
-    algolia?: unknown; // TODO type plugin
-  };
+export type ThemeConfig = BaseThemeConfig &
+  ClassicThemeConfig &
+  AlgoliaThemeConfig;
