+++ packages/docusaurus-theme-search-algolia/package.json
@@ -36,7 +36,8 @@
     "clsx": "^1.1.1",
     "eta": "^1.12.3",
     "lodash": "^4.17.20",
-    "tslib": "^2.3.1"
+    "tslib": "^2.3.1",
+    "utility-types": "^3.10.0"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "2.0.0-beta.14",
